+++ cms/src/layouts/DefaultLayout.tsx
@@ -1,13 +1,16 @@
 import React from "react";
+import { Outlet } from "react-router";
+
 import Header from "../components/header/Header";
 import Footer from "../components/footer/Footer";
-import Main from "../components/main/Main";
 
-const DefaultLayout: React.FC = () => {
+const DefaultLayout: React.FC<{ children?: React.ReactNode }> = ({
+  children,
+}) => {
   return (
     <>
       <Header />
-      <Main />
+      <main>{children ?? <Outlet />}</main>
       <Footer />
     </>
   );