+++ cms/src/types/errors.ts
@@ -1,0 +1,13 @@
+export interface LoginFormError {
+  username: ValidationError;
+  password: ValidationError;
+  message?: string;
+}
+
+interface ValidationError<InputType = string> {
+  location: string;
+  msg: string;
+  path: string;
+  type: string;
+  value: InputType;
+}